Use XpmNamespaceProvider for default namespaces.
