+++ src/lang-xpm/main/uk/co/reecedunn/intellij/plugin/xpm/optree/namespace/XpmNamespaceProvider.kt
@@ -26,4 +26,6 @@ interface XpmNamespaceProvider {
     }
 
     fun staticallyKnownNamespaces(context: PsiElement): Sequence<XpmNamespaceDeclaration>
+
+    fun defaultNamespace(context: PsiElement, type: XdmNamespaceType): Sequence<XpmNamespaceDeclaration>
 }
