+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/optree/XQueryNamespaceProvider.kt
@@ -18,11 +18,13 @@ package uk.co.reecedunn.intellij.plugin.xquery.optree
 import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.core.sequences.children
 import uk.co.reecedunn.intellij.plugin.core.sequences.walkTree
+import uk.co.reecedunn.intellij.plugin.xpm.optree.namespace.XdmNamespaceType
 import uk.co.reecedunn.intellij.plugin.xpm.optree.namespace.XpmNamespaceDeclaration
 import uk.co.reecedunn.intellij.plugin.xpm.optree.namespace.XpmNamespaceProvider
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryDirElemConstructor
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryModule
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryProlog
+import uk.co.reecedunn.intellij.plugin.xquery.model.defaultNamespace
 
 object XQueryNamespaceProvider : XpmNamespaceProvider {
     override fun staticallyKnownNamespaces(context: PsiElement): Sequence<XpmNamespaceDeclaration> {
@@ -38,5 +40,10 @@ object XQueryNamespaceProvider : XpmNamespaceProvider {
                 else -> emptySequence()
             }
         }.filterNotNull().distinct().filter { node -> node.namespacePrefix != null && node.namespaceUri != null }
+    }
+
+    override fun defaultNamespace(context: PsiElement, type: XdmNamespaceType): Sequence<XpmNamespaceDeclaration> {
+        if (context.containingFile !is XQueryModule) return emptySequence()
+        return context.defaultNamespace(type, true)
     }
 }
